Lab2/django/requirements.txt
django==5.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
